data/dataset.py
import numpy as np
import torch
from PIL import Image


class ImageClassificationDataset(torch.utils.data.Dataset):
    def __init__(self, image_path, labels, transforms=None):
        self._image_path = image_path
        self._labels = labels
        self._transforms = transforms

    def __len__(self):
        return len(self._labels)

    def __getitem__(self, index):
        label = torch.tensor(np.float64(self._labels[index]))
        image = Image.open(self._image_path[index]).convert('RGB')
        if self._transforms is not None:
            image = self._transforms(image)

        return image, label
