requirements.txt
pytorch-lightning==0.6.0
test-tube==0.7.1
torch==1.2.0
torchvision==0.4.0
numpy==1.17.2
tqdm==4.36.1
pretrainedmodels==0.7.4
Pillow==6.1.0
fire==0.2.1
scipy==1.3.1
tensorboardX==1.9
pytorch-ignite
